Refactor: Extract TraversalMixin (PR 2)
